requirements.txt
asgiref==3.4.1
astroid==2.7.1
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.8.0
django-db-logger==0.1.10
djangorestframework==3.12.4
google-auth==2.0.0
idna==3.2
isort==5.9.3
jsonfield==3.1.0
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.9.6
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
whitenoise==5.3.0
wrapt==1.12.1
